update modal api
