+++ components/Modal/common/contexts.ts
@@ -1,5 +1,8 @@
 import React from "react"
 
+/**
+ * Supplies common style attributes
+ */
 export const StyleContext = React.createContext({
   padding: {
     vertical: "2.1rem",
