+++ components/Modal/variants/BottomPanel.tsx
@@ -1,7 +1,6 @@
-import React, { ReactNode } from "react"
+import React from "react"
 import styled from "styled-components"
 
-import withModalWrapper from "./withModalWrapper"
 import { BaseLayout } from './Base'
 import { border } from "../common/styles"
 import { PanelModalProps } from "./interfaces"
@@ -28,4 +27,4 @@ const BottomPanelModal: React.FC<PanelModalProps> = ({
   )
 }
 
-export default withModalWrapper(BottomPanelModal)
+export default BottomPanelModal
