+++ components/Modal/variants/RightPanel.tsx
@@ -1,7 +1,6 @@
-import React, { ReactNode } from "react"
+import React from "react"
 import styled from "styled-components"
 
-import withModalWrapper from "./withModalWrapper"
 import { BaseLayout } from './Base'
 import { border } from "../common/styles"
 import { PanelModalProps } from "./interfaces"
@@ -45,4 +44,4 @@ const RightPanelModal: React.FC<PanelModalProps> = ({
   )
 }
 
-export default withModalWrapper(RightPanelModal)
+export default RightPanelModal
