+++ package.json
@@ -34,25 +34,6 @@
     "react": ">=16.0.0",
     "react-dom": ">=16.0.0"
   },
-  "jest": {
-    "transform": {
-      "^.+\\.jsx?$": "babel-jest",
-      "^.+\\.tsx?$": "ts-jest"
-    },
-    "moduleNameMapper": {
-      "\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/components/_utils/__mocks__/fileMock.ts",
-      "\\.(css|less)$": "identity-obj-proxy"
-    },
-    "preset": "ts-jest",
-    "globals": {
-      "ts-jest": {
-        "tsConfig": "./tsconfig.test.json"
-      }
-    },
-    "setupFilesAfterEnv": [
-      "react-testing-library/cleanup-after-each"
-    ]
-  },
   "devDependencies": {
     "@babel/cli": "^7.4.3",
     "@babel/core": "^7.4.3",
